pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aura9"
version = "2.1.0"
description = "Aura-9: Autonomous Reasoning Agent, Generation 9"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27,<1",
    "pynvml>=12.0,<13",
    "redis>=5.0,<6",
    "qdrant-client>=1.12,<2",
    "falkordb>=1.0,<2",
    "loguru>=0.7,<1",
    "pyyaml>=6.0,<7",
    "click>=8.1,<9",
    "uvicorn>=0.30,<1",
    "starlette>=0.40,<1",
    "prometheus-client>=0.21,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.24,<1",
    "ruff>=0.6,<1",
]

[project.scripts]
aura9 = "aura9.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/aura9"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B"]
ignore = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
